Add JLCPCB Order Number placeholder
This adds silkscreen under the MCU saying "JLCJLCJLCJLC". When ordering
with JLCPCB, they'll replace this with the order number. If this isn't
done, the order number will be placed somewhere random where it can
potentially be seen after assembly.

Since this is purely for a slight convenience when using JLCPCB, I can
also see just not doing this.

Also see:
https://support.jlcpcb.com/article/28-how-to-put-jlc-production-id-at-a-specified-area-on-the-pcb